separating classes and objects into different files from `TileProjectBench`
